3rd release match build release on play version 3
