+++ pubspec.yaml
@@ -52,6 +52,7 @@ dependencies:
   google_mobile_ads: ^1.2.0
   flutter_spinkit: ^5.1.0
   basic_utils: ^4.2.2
+  animated_text_kit: ^4.2.1
 dev_dependencies:
   flutter_test:
     sdk: flutter
